scripts/pyproject.toml
[project]
name = "gitlab-sync"
version = "0.1.0"
description = "A script to sync GitLab issues to local markdown files."
dependencies = [
    "python-gitlab",
    "pyyaml",
    "typer",
    "python-dotenv",
    "networkx"
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-mock",
]

[tool.pytest.ini_options]
pythonpath = ["."]

